Safari-611.1.10/LayoutTests/http/tests/misc/frame-default-enc-different-domain-expected.txt

ALERT: windows-1252
